remove pymupdf, replace with pdf2image
