+++ examples/gaia_agent/scripts/install.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+conda install conda-forge::poppler # for pdf2image to convert pdf pages to images
+pip install -r ../../../requirements.converters.txt
